[justin] added justin and adjusted transporting demo
