+++ binder/01-import.py
@@ -3,7 +3,7 @@ import sys
 sys.path.insert(0, '/home/jovyan/workspace/ros/src/pycram')
 
 from demos.pycram_virtual_building_demos.setup.setup_launch_robot import launch_pr2, launch_hsrb, launch_stretch, \
-    launch_tiago
+    launch_tiago, launch_justin
 from pycram.datastructures.enums import ObjectType, WorldMode
 from pycram.designators.object_designator import *
 from pycram.object_descriptors.urdf import ObjectDescription
@@ -23,3 +23,5 @@ elif robot_param == 'stretch':
     launch_stretch()
 elif robot_param == 'tiago':
     launch_tiago()
+elif robot_param == 'justin':
+    launch_justin()
