+++ demos/pycram_virtual_building_demos/setup_demo_manager.py
@@ -3,10 +3,11 @@ import logging
 from IPython.core.display_functions import clear_output
 from ipywidgets import Output
 
-from demos.pycram_bullet_world_demo.demo import transporting_demo
 from demos.pycram_virtual_building_demos.setup.setup_launch_robot import *
 import threading
 import sys
+
+from demos.pycram_virtual_building_demos.src.transport_demo import transporting_demo
 
 sys.path.insert(0, '/home/vee/robocup_workspaces/pycram_ws/src/pycram')
 
@@ -14,7 +15,8 @@ from ipywidgets import Output
 
 from demos.pycram_virtual_building_demos.setup.setup_launch_robot import launch_pr2, launch_hsrb, launch_stretch, \
     launch_tiago
-from demos.pycram_virtual_building_demos.setup.setup_utils import display_loading_gif_with_text, update_text
+from demos.pycram_virtual_building_demos.setup.setup_utils import display_loading_gif_with_text, update_text, \
+    get_robot_name
 import rospy
 
 from demos.pycram_virtual_building_demos.src.simple_examples import navigate_simple_example
@@ -53,6 +55,8 @@ def start_demo():
     robot_param = rospy.get_param('/nbparam_robots')
     task_param = rospy.get_param('/nbparam_tasks')
 
+    robot_name = get_robot_name(robot_param)
+
     extension = ObjectDescription.get_file_extension()
     # text widget for the virtual building
     text_widget = display_loading_gif_with_text()
@@ -60,24 +64,24 @@ def start_demo():
 
     world = BulletWorld(WorldMode.DIRECT)
     VizMarkerPublisher()
-    Object(robot_param, ObjectType.ROBOT, f"{robot_param}{extension}", pose=Pose([1, 2, 0]))
+    robot = Object(robot_name, ObjectType.ROBOT, f"{robot_name}{extension}", pose=Pose([1, 2, 0]))
     apartment = Object(environment_param, ObjectType.ENVIRONMENT, f"{environment_param}{extension}")
     tf = TFBroadcaster()
 
     update_text(text_widget, 'Executing Demo: ' + task_param)
 
-    demo_selecting(apartment, task_param)
+    demo_selecting(apartment, robot, task_param)
 
     extension = ObjectDescription.get_file_extension()
 
     update_text(text_widget, 'Done with: ' + task_param)
 
 
-def demo_selecting(apartment, task_param):
-    display(output)
-    with output:
-        if task_param == "navigate":
-            navigate_simple_example()
-        elif task_param == "transport":
-            rospy.loginfo('Starting transporting demo...')
-            transporting_demo(apartment)
+def demo_selecting(apartment, robot, task_param):
+    # display(output)
+    # with output:
+    if task_param == "navigate":
+        navigate_simple_example()
+    elif task_param == "transport":
+        rospy.loginfo('Starting transporting demo...')
+        transporting_demo(apartment, robot)
